+++ .gitignore
@@ -123,8 +123,7 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-# RPDK file
-richard-cloud9-environment.zip
-#cloudformation-cli-python-lib-0.0.1.tar.gz
-service-2.json
-docs/
+# contains credentials
+sam-tests/
+
+rpdk.log*
