+++ .rpdk-config
@@ -1,10 +1,18 @@
 {
-    "typeName": "Richard::Cloud9::Environment",
+    "artifact_type": "RESOURCE",
+    "typeName": "Richard::Cloud9::CustomEC2",
     "language": "python37",
     "runtime": "python3.7",
-    "entrypoint": "richard_cloud9_environment.handlers.resource",
-    "testEntrypoint": "richard_cloud9_environment.handlers.test_entrypoint",
+    "entrypoint": "richard_cloud9_customec2.handlers.resource",
+    "testEntrypoint": "richard_cloud9_customec2.handlers.test_entrypoint",
     "settings": {
-        "use_docker": false
+        "version": false,
+        "subparser_name": null,
+        "verbose": 0,
+        "force": false,
+        "type_name": null,
+        "artifact_type": null,
+        "use_docker": true,
+        "protocolVersion": "2.0.0"
     }
 }
