+++ docs/README.md
@@ -1,0 +1,149 @@
+# Richard::Cloud9::CustomEC2
+
+An example resource schema demonstrating some basic constructs and validation rules.
+
+## Syntax
+
+To declare this entity in your AWS CloudFormation template, use the following syntax:
+
+### JSON
+
+<pre>
+{
+    "Type" : "Richard::Cloud9::CustomEC2",
+    "Properties" : {
+        "<a href="#name" title="Name">Name</a>" : <i>String</i>,
+        "<a href="#description" title="Description">Description</a>" : <i>String</i>,
+        "<a href="#instancetype" title="InstanceType">InstanceType</a>" : <i>String</i>,
+        "<a href="#subnetid" title="SubnetId">SubnetId</a>" : <i>String</i>,
+        "<a href="#operatingsystem" title="OperatingSystem">OperatingSystem</a>" : <i>String</i>,
+        "<a href="#idletimeout" title="IdleTimeout">IdleTimeout</a>" : <i>String</i>,
+        "<a href="#owner" title="Owner">Owner</a>" : <i>String</i>,
+        "<a href="#includes" title="Includes">Includes</a>" : <i>[ String, ... ]</i>,
+        "<a href="#tags" title="Tags">Tags</a>" : <i>[ <a href="tag.md">Tag</a>, ... ]</i>
+    }
+}
+</pre>
+
+### YAML
+
+<pre>
+Type: Richard::Cloud9::CustomEC2
+Properties:
+    <a href="#name" title="Name">Name</a>: <i>String</i>
+    <a href="#description" title="Description">Description</a>: <i>String</i>
+    <a href="#instancetype" title="InstanceType">InstanceType</a>: <i>String</i>
+    <a href="#subnetid" title="SubnetId">SubnetId</a>: <i>String</i>
+    <a href="#operatingsystem" title="OperatingSystem">OperatingSystem</a>: <i>String</i>
+    <a href="#idletimeout" title="IdleTimeout">IdleTimeout</a>: <i>String</i>
+    <a href="#owner" title="Owner">Owner</a>: <i>String</i>
+    <a href="#includes" title="Includes">Includes</a>: <i>
+      - String</i>
+    <a href="#tags" title="Tags">Tags</a>: <i>
+      - <a href="tag.md">Tag</a></i>
+</pre>
+
+## Properties
+
+#### Name
+
+A TPS Code is automatically generated on creation and assigned as the unique identifier.
+
+_Required_: No
+
+_Type_: String
+
+_Pattern_: <code>^[A-Z]{3,5}[0-9]{8}-[0-9]{4}$</code>
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### Description
+
+The title of the TPS report is a mandatory element.
+
+_Required_: No
+
+_Type_: String
+
+_Minimum_: <code>20</code>
+
+_Maximum_: <code>250</code>
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### InstanceType
+
+_Required_: Yes
+
+_Type_: String
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### SubnetId
+
+_Required_: No
+
+_Type_: String
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### OperatingSystem
+
+_Required_: No
+
+_Type_: String
+
+_Allowed Values_: <code>AMAZON_LINUX_2</code> | <code>AMAZON_LINUX_2012</code> | <code>UBUNTU_18_04</code>
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### IdleTimeout
+
+_Required_: No
+
+_Type_: String
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### Owner
+
+_Required_: No
+
+_Type_: String
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### Includes
+
+_Required_: No
+
+_Type_: List of String
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+#### Tags
+
+An array of key-value pairs to apply to this resource.
+
+_Required_: No
+
+_Type_: List of <a href="tag.md">Tag</a>
+
+_Update requires_: [No interruption](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-no-interrupt)
+
+## Return Values
+
+### Ref
+
+When you pass the logical ID of this resource to the intrinsic `Ref` function, Ref returns the Arn.
+
+### Fn::GetAtt
+
+The `Fn::GetAtt` intrinsic function returns a value for a specified attribute of this type. The following are the available attributes and sample return values.
+
+For more information about using the `Fn::GetAtt` intrinsic function, see [Fn::GetAtt](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-getatt.html).
+
+#### Arn
+
+#### EnvironmentId
+
