--- events/hasEnvId.json
@@ -1,21 +1,0 @@
-{
-  "credentials": {
-    "accessKeyId": "",
-    "secretAccessKey": "",
-    "sessionToken": ""
-  },
-  "action": "CREATE",
-  "request": {
-    "clientRequestToken": "ecba020e-b2e6-4742-a7d0-8a06ae7c4b2b",
-    "desiredResourceState": {
-      "InstanceType": "t2.micro",
-      "EBSVolumeSize": "50"
-    },
-    "previousResourceState": null,
-    "logicalResourceIdentifier": "MyLogicalId002"
-  },
-  "callbackContext": {
-    "ENVIRONMENT_ID": "d8a917e9a5bf42f7938368579e577893",
-    "LOCAL_STATUS":"ENVIRONMENT_CREATED"
-  }
-}
