--- events/insatceStopped.json
@@ -1,23 +1,0 @@
-{
-  "credentials": {
-    "accessKeyId": "",
-    "secretAccessKey": "",
-    "sessionToken": ""
-  },
-  "action": "CREATE",
-  "request": {
-    "clientRequestToken": "ecba020e-b2e6-4742-a7d0-8a06ae7c4b2b",
-    "desiredResourceState": {
-      "InstanceType": "t2.micro",
-      "EBSVolumeSize": 50,
-      "InstanceId":"i-08563fafd0bd5dd88",
-      "UserData": "VkVSU0lPTj0zLjcuNAp5dW0gdXBkYXRlIC15Cnl1bSBpbnN0YWxsIGdjYyBvcGVuc3NsLWRldmVsIGJ6aXAyLWRldmVsIGxpYmZmaS1kZXZlbCAteQp3Z2V0IGh0dHBzOi8vd3d3LnB5dGhvbi5vcmcvZnRwL3B5dGhvbi8ke1ZFUlNJT059L1B5dGhvbi0ke1ZFUlNJT059LnRnegp0YXIgeHpmIFB5dGhvbi0ke1ZFUlNJT059LnRnegpjZCBQeXRob24tJHtWRVJTSU9OfQouL2NvbmZpZ3VyZSAtLWVuYWJsZS1vcHRpbWl6YXRpb25zCm1ha2UgYWx0aW5zdGFsbApjZCAuLi8KIyBSZW1vdmUgb2xkIHN5bWxpbmtzCnJtIC1yZiAvZXRjL2FsdGVybmF0aXZlcy9waXAKcm0gLXJmIC9ldGMvYWx0ZXJuYXRpdmVzL3B5dGhvbgojIG1ha2UgbmV3IHN5bWxpbmtzCmxuIC1zIC91c3IvbG9jYWwvYmluL3BpcCR7VkVSU0lPTjowOjN9IC9ldGMvYWx0ZXJuYXRpdmVzL3BpcApsbiAtcyAvdXNyL2xvY2FsL2Jpbi9weXRob24ke1ZFUlNJT046MDozfSAvZXRjL2FsdGVybmF0aXZlcy9weXRob24K"
-    },
-    "previousResourceState": null,
-    "logicalResourceIdentifier": "MyLogicalId002"
-  },
-  "callbackContext":{
-    "ENVIRONMENT_ID":"d8a917e9a5bf42f7938368579e577893",
-    "LOCAL_STATUS":"STOPPED_INSTANCE"
-  }
-}
