--- events/integ-final.yaml
@@ -1,21 +1,0 @@
-AWSTemplateFormatVersion: 2010-09-09
-Resources:
-  IAMTest:
-    Type: Richard::Cloud9::Environment
-    Properties:
-      InstanceType: c5.large
-      EBSVolumeSize: 50
-      OwnerArn: arn:aws:sts::784907117618:assumed-role/CrossAccountRole-Isengard/rhboyd-Isengard
-      Cloud9InstancePolicy:
-        PolicyName: "MyInstancePolicy"
-        PolicyDocument:
-          Version: '2012-10-17'
-          Statement:
-            - Effect: Allow
-              Action:
-                - "*"
-              Resource: 
-                - "*"
-      UserData:
-        Bucket: rhb-blog
-        Object: userdata/userdata.sh
