--- events/local.json
@@ -1,23 +1,0 @@
-{
-    "desiredResourceState": {
-      "InstanceType": "t2.micro",
-      "EBSVolumeSize": "50",
-      "OwnerArn": "arn:aws:sts::072954315904:assumed-role/CrossAccountRole-Isengard/rhboyd-Isengard",
-      "Cloud9InstancePolicy": {
-        "PolicyName": "MyInstancePolicy",
-        "PolicyDocument": {
-          "Version": "2012-10-17",
-          "Statement":[{
-            "Effect": "Allow",
-            "Action":["*"],
-            "Resource": ["*"]
-          }]
-        }
-      },
-      "UserData": {
-        "Bucket": "cloudformationmanageduploadinfrast-artifactbucket-1omv9gdsrde06",
-        "Object": "userdata.sh"
-      }
-    },
-    "logicalResourceIdentifier": "shortUserData008"
-}
