+++ example_inputs/inputs_1_create.json
@@ -1,0 +1,12 @@
+{
+    "TPSCode": "...",
+    "Title": "...",
+    "CoverSheetIncluded": "...",
+    "DueDate": "...",
+    "ApprovalDate": "...",
+    "Memo": "...",
+    "SecondCopyOfMemo": "...",
+    "TestCode": "...",
+    "Authors": "...",
+    "Tags": "..."
+}
