+++ requirements.txt
@@ -1,2 +1,1 @@
-cloudformation-cli-python-lib==0.0.1
-requests
+cloudformation-cli-python-lib>=2.1.3
