+++ resource-role.yaml
@@ -23,7 +23,11 @@ Resources:
             Statement:
               - Effect: Allow
                 Action:
-                - "*"
+                - "initech:CreateReport"
+                - "initech:DeleteReport"
+                - "initech:DescribeReport"
+                - "initech:ListReports"
+                - "initech:UpdateReport"
                 Resource: "*"
 Outputs:
   ExecutionRoleArn:
