+++ src/richard_cloud9_customec2/models.py
@@ -1,0 +1,122 @@
+# DO NOT modify this file by hand, changes will be overwritten
+import sys
+from dataclasses import dataclass
+from inspect import getmembers, isclass
+from typing import (
+    AbstractSet,
+    Any,
+    Generic,
+    Mapping,
+    MutableMapping,
+    Optional,
+    Sequence,
+    Type,
+    TypeVar,
+)
+
+from cloudformation_cli_python_lib.interface import (
+    BaseModel,
+    BaseResourceHandlerRequest,
+)
+from cloudformation_cli_python_lib.recast import recast_object
+from cloudformation_cli_python_lib.utils import deserialize_list
+
+T = TypeVar("T")
+
+
+def set_or_none(value: Optional[Sequence[T]]) -> Optional[AbstractSet[T]]:
+    if value:
+        return set(value)
+    return None
+
+
+@dataclass
+class ResourceHandlerRequest(BaseResourceHandlerRequest):
+    # pylint: disable=invalid-name
+    desiredResourceState: Optional["ResourceModel"]
+    previousResourceState: Optional["ResourceModel"]
+    typeConfiguration: Optional["TypeConfigurationModel"]
+
+
+@dataclass
+class ResourceModel(BaseModel):
+    Name: Optional[str]
+    Description: Optional[str]
+    InstanceType: Optional[str]
+    SubnetId: Optional[str]
+    OperatingSystem: Optional[str]
+    IdleTimeout: Optional[str]
+    Owner: Optional[str]
+    Arn: Optional[str]
+    EnvironmentId: Optional[str]
+    Includes: Optional[Sequence[str]]
+    Tags: Optional[AbstractSet["_Tag"]]
+
+    @classmethod
+    def _deserialize(
+        cls: Type["_ResourceModel"],
+        json_data: Optional[Mapping[str, Any]],
+    ) -> Optional["_ResourceModel"]:
+        if not json_data:
+            return None
+        dataclasses = {n: o for n, o in getmembers(sys.modules[__name__]) if isclass(o)}
+        recast_object(cls, json_data, dataclasses)
+        return cls(
+            Name=json_data.get("Name"),
+            Description=json_data.get("Description"),
+            InstanceType=json_data.get("InstanceType"),
+            SubnetId=json_data.get("SubnetId"),
+            OperatingSystem=json_data.get("OperatingSystem"),
+            IdleTimeout=json_data.get("IdleTimeout"),
+            Owner=json_data.get("Owner"),
+            Arn=json_data.get("Arn"),
+            EnvironmentId=json_data.get("EnvironmentId"),
+            Includes=json_data.get("Includes"),
+            Tags=set_or_none(json_data.get("Tags")),
+        )
+
+
+# work around possible type aliasing issues when variable has same name as a model
+_ResourceModel = ResourceModel
+
+
+@dataclass
+class Tag(BaseModel):
+    Key: Optional[str]
+    Value: Optional[str]
+
+    @classmethod
+    def _deserialize(
+        cls: Type["_Tag"],
+        json_data: Optional[Mapping[str, Any]],
+    ) -> Optional["_Tag"]:
+        if not json_data:
+            return None
+        return cls(
+            Key=json_data.get("Key"),
+            Value=json_data.get("Value"),
+        )
+
+
+# work around possible type aliasing issues when variable has same name as a model
+_Tag = Tag
+
+
+@dataclass
+class TypeConfigurationModel(BaseModel):
+
+    @classmethod
+    def _deserialize(
+        cls: Type["_TypeConfigurationModel"],
+        json_data: Optional[Mapping[str, Any]],
+    ) -> Optional["_TypeConfigurationModel"]:
+        if not json_data:
+            return None
+        return cls(
+        )
+
+
+# work around possible type aliasing issues when variable has same name as a model
+_TypeConfigurationModel = TypeConfigurationModel
+
+
