--- tests/context.py
@@ -1,6 +1,0 @@
-# context.py
-import os
-import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../src/richard_cloud9_environment')))
-
-import interface
